pyproject.toml
[project]
name = "maint-bot-app"
version = "0.1.0"
description = "MaintBot — Manufacturing Maintenance Assistant with Lakebase persistence"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "uvicorn>=0.41.0",
    "databricks-langchain[memory]>=0.17.0",
    "databricks-sdk>=0.79.0",
    "databricks-agents>=1.9.3",
    "mlflow>=3.10.0",
    "langgraph>=1.0.9",
    "langchain-mcp-adapters>=0.2.1",
    "python-dotenv>=1.2.1",
    "tiktoken",
    "psycopg[binary]",
    "numpy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent_server", "scripts"]

[dependency-groups]
dev = [
    "hatchling>=1.28.0",
    "pytest>=9.0.2",
]

[project.scripts]
quickstart = "scripts.quickstart:main"
start-app = "scripts.start_app:main"
start-server = "agent_server.start_server:main"
